Extract profile folding logic
as it doesn't quite belong in the entrypoint.
